Tests: disable TestDatastoreSuite when race detector enabled
Move the TestDatastoreSuite to a file with // +build !race so that it does not
run when the race detector is enabled.
